Consolidate test_eqjac_gpu and test_eqjac_perf.
